+++ src/main/java/com/soongsil/CoffeeChat/config/AsyncConfig.java
@@ -1,4 +1,4 @@
-package com.soongsil.CoffeeChat.config.async;
+package com.soongsil.CoffeeChat.config;
 
 import java.util.concurrent.Executor;
