+++ src/main/java/com/soongsil/CoffeeChat/controller/ReportController.java
@@ -1,6 +1,6 @@
 package com.soongsil.CoffeeChat.controller;
 
-import com.soongsil.CoffeeChat.dto.Oauth.CustomOAuth2User;
+import com.soongsil.CoffeeChat.security.dto.CustomOAuth2User;
 import com.soongsil.CoffeeChat.dto.ReportDto;
 import com.soongsil.CoffeeChat.service.ReportService;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
@@ -8,7 +8,6 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.Authentication;
-import org.springframework.web.ErrorResponse;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
