--- src/main/java/com/soongsil/CoffeeChat/dto/Oauth/GoogleResponse.java
@@ -1,36 +1,0 @@
-package com.soongsil.CoffeeChat.dto.Oauth;
-
-import java.util.Map;
-
-public class GoogleResponse implements OAuth2Response {
-	private final Map<String, Object> attribute;
-
-	public GoogleResponse(Map<String, Object> attribute) {
-
-		this.attribute = attribute;
-	}
-
-	@Override
-	public String getProvider() {
-
-		return "google";
-	}
-
-	@Override
-	public String getProviderId() {
-
-		return attribute.get("sub").toString();
-	}
-
-	@Override
-	public String getEmail() {
-
-		return attribute.get("email").toString();
-	}
-
-	@Override
-	public String getName() {
-
-		return attribute.get("name").toString();
-	}
-}
