+++ src/main/java/com/soongsil/CoffeeChat/security/dto/UserDTO.java
@@ -1,0 +1,12 @@
+package com.soongsil.CoffeeChat.security.dto;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class UserDTO {
+	private String role;
+	private String name;
+	private String username; //스프링 서버 내의 유저 아이디
+}
