+++ src/main/java/com/soongsil/CoffeeChat/security/jwt/CustomLogoutFilter.java
@@ -1,0 +1,111 @@
+package com.soongsil.CoffeeChat.security.jwt;
+
+import java.io.IOException;
+
+import org.springframework.web.filter.GenericFilterBean;
+
+import com.soongsil.CoffeeChat.repository.RefreshRepository;
+
+import io.jsonwebtoken.ExpiredJwtException;
+import jakarta.servlet.FilterChain;
+import jakarta.servlet.ServletException;
+import jakarta.servlet.ServletRequest;
+import jakarta.servlet.ServletResponse;
+import jakarta.servlet.http.Cookie;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+
+public class CustomLogoutFilter extends GenericFilterBean {
+	private final JWTUtil jwtUtil;
+	private final RefreshRepository refreshRepository;
+
+	public CustomLogoutFilter(JWTUtil jwtUtil, RefreshRepository refreshRepository) {
+
+		this.jwtUtil = jwtUtil;
+		this.refreshRepository = refreshRepository;
+	}
+
+	@Override
+	public void doFilter(ServletRequest request, ServletResponse response, FilterChain chain) throws
+		IOException,
+		ServletException {
+		doFilter((HttpServletRequest)request, (HttpServletResponse)response, chain);
+	}
+
+	private void doFilter(HttpServletRequest request, HttpServletResponse response, FilterChain filterChain) throws
+		IOException,
+		ServletException {
+
+		//path and method verify
+		String requestUri = request.getRequestURI();  //요청의 path값을 꺼내고
+		if (!requestUri.matches("^\\/logout$")) {  //logout경로인지 확인 후
+
+			filterChain.doFilter(request, response);
+			return;
+		}
+		String requestMethod = request.getMethod();
+		if (!requestMethod.equals("POST")) {        //POST요청인지 확인
+
+			filterChain.doFilter(request, response);
+			return;
+		}
+
+		//get refresh token
+		String refresh = null;
+		Cookie[] cookies = request.getCookies();
+		for (Cookie cookie : cookies) {
+
+			if (cookie.getName().equals("refresh")) {
+
+				refresh = cookie.getValue();
+			}
+		}
+
+		//refresh null check 400 (refresh토큰이 null상태)
+		if (refresh == null) {
+
+			response.setStatus(HttpServletResponse.SC_BAD_REQUEST);
+			return;
+		}
+
+		//expired check
+		try {
+			jwtUtil.isExpired(refresh);
+		} catch (ExpiredJwtException e) {
+
+			//response status code 400 (만료된 토큰)
+			response.setStatus(HttpServletResponse.SC_BAD_REQUEST);
+			return;
+		}
+
+		// 토큰이 refresh인지 확인 (발급시 페이로드에 명시)
+		String category = jwtUtil.getCategory(refresh);
+		if (!category.equals("refresh")) {
+
+			//response status code 400 (refresh토큰이 아님)
+			response.setStatus(HttpServletResponse.SC_BAD_REQUEST);
+			return;
+		}
+
+		//DB에 저장되어 있는지 확인
+		Boolean isExist = refreshRepository.existsByRefresh(refresh);
+		if (!isExist) {
+
+			//response status code 400 (DB에 저장된 refresh 토큰이 아님)
+			response.setStatus(HttpServletResponse.SC_BAD_REQUEST);
+			return;
+		}
+
+		//로그아웃 진행
+		//Refresh 토큰 DB에서 제거 (앞으로 해당 refresh토큰으로 reissue불가능)
+		refreshRepository.deleteByRefresh(refresh);
+
+		//Refresh 토큰 Cookie 값 0  (null인 refresh토큰 반환)
+		Cookie cookie = new Cookie("refresh", null);
+		cookie.setMaxAge(0);
+		cookie.setPath("/");
+
+		response.addCookie(cookie);
+		response.setStatus(HttpServletResponse.SC_OK);
+	}
+}
